test: merge NamedTuple and class tests
